data: Fix NullPointerException when downloading titles
GameTitles initializes every HiveMode once to get its ID.
Currently, initializing a HiveMode requires Beezig to have
a TemporaryPointsManager which is initialized after the
data is updated; hence Beezig wasn't able to update
GameTitles as Beezig.get().getTemporaryPointsManager()
returns null. This patch changes HiveMode so that it
doesn't initialize dailyService if no
temporaryPointsManager is available in Beezig.
This could lead to unpredictable results, but currently,
the fresh instances are just used to return their ID.

Signed-off-by: Archer <archer@beezig.eu>